Fix list and a few home tests
