+++ cypress.config.ts
@@ -17,7 +17,7 @@ export default defineConfig({
     // We've imported your old cypress plugins here.
     // You may want to clean this up later by importing these.
     setupNodeEvents(on, config) {
-      return require('./cypress/plugins/index.js')(on, config);
+      return require('./cypress/plugins/index.ts')(on, config);
     },
     baseUrl: 'http://localhost:3030',
     specPattern: 'cypress/e2e/**/*.{js,jsx,ts,tsx}',
